pyproject.toml
[project]
name = "resonate-async-http-api-ps"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.116.1",
    "resonate-sdk>=0.6.2",
    "uvicorn>=0.35.0",
]
